movi a função "agruparGenero"  que estava igual em dois js diferentes para utils.js e apenas importei nos js que necessitavam
